+++ net/ghttp/ghttp_server_service_handler.go
@@ -293,8 +293,8 @@ func createRouterFunc(funcInfo handlerFuncInfo) func(r *Request) {
 			inputValues = append(inputValues, inputObject)
 		}
 		if runtime.GOARCH == "amd64" {
-			if funcInfo.handlerFuncClosure != nil {
-				asmCall(&funcInfo, r, reqParameter)
+			if funcInfo.IsStrictRoute && funcInfo.handlerFuncClosure != nil {
+				asmCallStrictRoute(&funcInfo, r, reqParameter)
 				return
 			}
 		}
@@ -318,47 +318,6 @@ func createRouterFunc(funcInfo handlerFuncInfo) func(r *Request) {
 		}
 	}
 }
-
-func asmCall(funcInfo *handlerFuncInfo, r *Request, req unsafe.Pointer) {
-	type iface struct {
-		typ  unsafe.Pointer
-		data unsafe.Pointer
-	}
-	switch funcInfo.Type.NumOut() {
-	case 2:
-		if funcInfo.Type.Out(0).Kind() == reflect.Slice {
-			res, err := doAnyCallRequest_with_sliceRes_err(funcInfo.handlerFuncClosure, r.Context(), req)
-			sliceValue := reflect.New(funcInfo.Type.Out(0)).Elem().Interface()
-			out := (*iface)(unsafe.Pointer(&sliceValue))
-			(*out).data = unsafe.Pointer(&res)
-			r.handlerResponse = sliceValue
-			r.error = err
-		} else {
-			res, err := doAnyCallRequest_with_res_err(funcInfo.handlerFuncClosure, r.Context(), req)
-			outType := funcInfo.Type.Out(0).Elem()
-			outValue := reflect.New(outType).Interface()
-			out := (*iface)(unsafe.Pointer(&outValue))
-			(*out).data = res
-			r.handlerResponse = outValue
-			r.error = err
-		}
-	case 1:
-		err := doAnyCallRequest_with_err(funcInfo.handlerFuncClosure, r.Context(), req)
-		r.error = err
-	}
-}
-
-func doAnyCallRequest_with_err(fn any, ctx context.Context, req unsafe.Pointer) error
-
-func doAnyCallRequest_with_res_err(fn any, ctx context.Context, req unsafe.Pointer) (unsafe.Pointer, error)
-
-type _slice struct {
-	ptr unsafe.Pointer
-	len int
-	cap int
-}
-
-func doAnyCallRequest_with_sliceRes_err(fn any, ctx context.Context, req unsafe.Pointer) (_slice, error)
 
 // trimGeneric removes type definitions string from response type name if generic
 func trimGeneric(structName string) string {
